fix: worker thread is not using the latest loader implementation
