+++ apps/engine/src/preload.ts
@@ -1,6 +1,6 @@
-import { createRequire, Module } from 'node:module';
-import { resolve } from 'node:path';
+import { Module } from 'node:module';
 import TachybaseGlobal from '@tachybase/globals';
+import { defineLoader } from '@tachybase/loader/loader';
 
 import { DEFAULT_ENGINE_PACKAGES_PATH, DEFAULT_ENGINE_PLUGIN_PATH } from './constants';
 
@@ -76,29 +76,4 @@ process.env.TACHYBASE_WORKER_PATHS = lookingPaths.join(',');
 process.env.TACHYBASE_WORKER_MODULES = [...whitelists].join(',');
 
 // 整个加载过程允许报错，保持和默认加载器一样的行为
-Module._load = function (request: string, parent: NodeModule | null, isMain: boolean) {
-  // 使用白名单拦截，以及所有符合 '@tachybase/' 前缀的包
-  if (whitelists.has(request) || request.startsWith('@tachybase/')) {
-    try {
-      const resolvedFromApp = require.resolve(request, { paths: lookingPaths });
-      return originalLoad(resolvedFromApp, parent, isMain);
-    } catch (err) {
-      for (const basePath of lookingPaths) {
-        // 支持非 node_modules 的加载
-        try {
-          const pluginRoot = resolve(basePath, request);
-          const fakeRequire = createRequire(pluginRoot + '/index.js');
-          const resolved = fakeRequire.resolve(pluginRoot);
-          return originalLoad(resolved, parent, isMain);
-        } catch {}
-      }
-      // 这里不应该发生，但是我们依旧提供回退的机制，使用默认行为来加载模块
-      if (err.code === 'MODULE_NOT_FOUND') {
-        return originalLoad(request, parent, isMain);
-      }
-    }
-  }
-
-  // 相对路径、绝对路径不动
-  return originalLoad(request, parent, isMain);
-};
+Module._load = defineLoader(whitelists, originalLoad, lookingPaths);
