+++ packages/loader/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@tachybase/loader",
+  "version": "1.3.5",
+  "description": "",
+  "license": "Apache-2.0",
+  "main": "./src/index.js",
+  "scripts": {
+    "build": "tachybase-build --no-dts @tachybase/devtools"
+  },
+  "dependencies": {
+    "fast-glob": "^3.3.2"
+  }
+}
