+++ packages/module-worker-thread/package.json
@@ -19,6 +19,7 @@
     "@tachybase/actions": "workspace:*",
     "@tachybase/client": "workspace:*",
     "@tachybase/globals": "workspace:*",
+    "@tachybase/loader": "workspace:*",
     "@tachybase/logger": "workspace:*",
     "@tachybase/module-collection": "workspace:*",
     "@tachybase/module-user": "workspace:*",
